Progress Track Full

+++ BaseCustomerMVC/Models/ChapterProgressViewModel.cs
@@ -10,7 +10,7 @@ namespace BaseCustomerMVC.Models
         [JsonProperty("CompletedLessons")]
         public int CompletedLessons { get; set; }
         [JsonProperty("TotalLessons")]
-        public int TotalLessons { get; set; }
+        public long TotalLessons { get; set; }
         [JsonProperty("LastLessonID")]
         public string LastLessonID { get; set; }
         [JsonProperty("LastDate")]
